requirements.txt
# Core dependencies
python>=3.10
gymnasium>=0.29.0
numpy>=1.24.0
torch>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Interface
streamlit>=1.28.0
streamlit-aggrid>=0.3.0

# Analysis
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.4.0
flake8>=6.0.0
isort>=5.12.0

# Utils
pyyaml>=6.0
tqdm>=4.65.0
